+++ napari/components/add_layers_mixin.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, List, Optional, Sequence, Union
 import numpy as np
 
 from .. import layers
-from ..plugins.io import read_data_with_plugins, write_layers_with_plugins
+from ..plugins.io import read_data_with_plugins
 from ..utils import colormaps, io
 from ..utils.misc import (
     ensure_iterable,
@@ -938,69 +938,6 @@ class AddLayersMixin:
                 raise exc
 
         return layer
-
-    def save_layers(
-        self, path: str, selected: bool = True, plugin: Optional[str] = None
-    ):
-        """Save all or only selected layers to a path using writer plugins.
-
-        If ``plugin`` is not provided and only one layer is targeted, then
-        we just directly call ``napari_write_<layer_type>`` hook specification
-        (see :ref:`single layer writer hookspecs
-        <write-single-layer-hookspecs>`) which will loop through
-        implementations and stop when the first one returns a non-``None``
-        result. The order in which implementations are called can be changed
-        with the Plugin sorter in the GUI or with the corresponding hook's
-        :meth:`~napari.plugins._hook_callers._HookCaller.bring_to_front`
-        method.
-
-        If ``plugin`` is not provided and multiple layers are targeted,
-        then we call
-        :meth:`~napari.plugins.hook_specifications.napari_get_writer` which
-        loops through plugins to find the first one that knows how to handle
-        the combination of layers and is able to write the file. If no plugins
-        offer :meth:`~napari.plugins.hook_specifications.napari_get_writer` for
-        that combination of layers then the default
-        :meth:`~napari.plugins.hook_specifications.napari_get_writer` will
-        create a folder and call ``napari_write_<layer_type>`` for each layer
-        using the ``Layer.name`` variable to modify the path such that the
-        layers are written to unique files in the folder.
-
-        If ``plugin`` is provided and a single layer is targeted, then we
-        call the ``napari_write_<layer_type>`` for that plugin, and if it fails
-        we error.
-
-        If ``plugin`` is provided and multiple layers are targeted, then
-        we call we call
-        :meth:`~napari.plugins.hook_specifications.napari_get_writer` for
-        that plugin, and if it doesn’t return a ``WriterFunction`` we error,
-        otherwise we call it and if that fails if it we error.
-
-        Parameters
-        ----------
-        path : str
-            A filepath, directory, or URL to open.  Extensions may be used to
-            specify output format (provided a plugin is avaiable for the
-            requested format).
-        selected : bool
-            Optional flag to only save selected layers. True by default.
-        plugin : str, optional
-            Name of the plugin to use for saving. If None then all plugins
-            corresponding to appropriate hook specification will be looped
-            through to find the first one that can save the data.
-        """
-        if selected:
-            layers = self.layers.selected  # type: ignore
-        else:
-            layers = list(self.layers)  # type: ignore
-
-        if not layers:
-            import warnings
-
-            warnings.warn(f"No layers {'selected' if selected else 'to save'}")
-            return
-
-        write_layers_with_plugins(path=path, layers=layers, plugin_name=plugin)
 
 
 def prune_kwargs(kwargs: Dict[str, Any], layer_type: str) -> Dict[str, Any]:
