+++ napari/components/layerlist.py
@@ -1,3 +1,4 @@
+from typing import Optional
 from ..layers import Layer
 from ..utils.naming import inc_name_count
 from ..utils.list import ListModel
@@ -184,3 +185,77 @@ class LayerList(ListModel):
         for layer in self:
             if layer.selected:
                 layer.visible = not layer.visible
+
+    def save(
+        self,
+        path: str,
+        *,
+        selected: bool = False,
+        plugin: Optional[str] = None,
+    ):
+        """Save all or only selected layers to a path using writer plugins.
+
+        If ``plugin`` is not provided and only one layer is targeted, then
+        we just directly call ``napari_write_<layer_type>`` hook specification
+        (see :ref:`single layer writer hookspecs
+        <write-single-layer-hookspecs>`) which will loop through
+        implementations and stop when the first one returns a non-``None``
+        result. The order in which implementations are called can be changed
+        with the Plugin sorter in the GUI or with the corresponding hook's
+        :meth:`~napari.plugins._hook_callers._HookCaller.bring_to_front`
+        method.
+
+        If ``plugin`` is not provided and multiple layers are targeted,
+        then we call
+        :meth:`~napari.plugins.hook_specifications.napari_get_writer` which
+        loops through plugins to find the first one that knows how to handle
+        the combination of layers and is able to write the file. If no plugins
+        offer :meth:`~napari.plugins.hook_specifications.napari_get_writer` for
+        that combination of layers then the default
+        :meth:`~napari.plugins.hook_specifications.napari_get_writer` will
+        create a folder and call ``napari_write_<layer_type>`` for each layer
+        using the ``Layer.name`` variable to modify the path such that the
+        layers are written to unique files in the folder.
+
+        If ``plugin`` is provided and a single layer is targeted, then we
+        call the ``napari_write_<layer_type>`` for that plugin, and if it fails
+        we error.
+
+        If ``plugin`` is provided and multiple layers are targeted, then
+        we call we call
+        :meth:`~napari.plugins.hook_specifications.napari_get_writer` for
+        that plugin, and if it doesn’t return a ``WriterFunction`` we error,
+        otherwise we call it and if that fails if it we error.
+
+        Parameters
+        ----------
+        path : str
+            A filepath, directory, or URL to open.  Extensions may be used to
+            specify output format (provided a plugin is avaiable for the
+            requested format).
+        selected : bool
+            Optional flag to only save selected layers. False by default.
+        plugin : str, optional
+            Name of the plugin to use for saving. If None then all plugins
+            corresponding to appropriate hook specification will be looped
+            through to find the first one that can save the data.
+
+        Returns
+        -------
+        bool
+            Return True if data is successfully written.
+        """
+        from ..plugins.io import save_layers
+
+        if selected:
+            layers = self.selected
+        else:
+            layers = list(self)
+
+        if not layers:
+            import warnings
+
+            warnings.warn(f"No layers {'selected' if selected else 'to save'}")
+            return
+
+        return save_layers(path, layers, plugin=plugin)
